# docs/api/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    OperationalExerciseCreate:
      description: Validated request for one bounded monitoring exercise.
      properties:
        duration_seconds:
          default: 240
          maximum: 1800.0
          minimum: 120.0
          title: Duration Seconds
          type: integer
        exercise_type:
          pattern: ^(api_unavailable|high_latency|elevated_5xx)$
          title: Exercise Type
          type: string
      required:
      - exercise_type
      title: OperationalExerciseCreate
      type: object
    OperationalExerciseResponse:
      additionalProperties: false
      description: One controlled monitoring exercise.
      properties:
        exercise_type:
          enum:
          - api_unavailable
          - high_latency
          - elevated_5xx
          title: Exercise Type
          type: string
        expires_at:
          title: Expires At
          type: string
        id:
          title: Id
          type: string
        initiated_by:
          title: Initiated By
          type: string
        recovered_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Recovered At
        started_at:
          title: Started At
          type: string
        status:
          anyOf:
          - enum:
            - active
            - recovered
            type: string
          - type: 'null'
          title: Status
      required:
      - id
      - exercise_type
      - initiated_by
      - started_at
      - expires_at
      title: OperationalExerciseResponse
      type: object
    OperationalExerciseStateResponse:
      additionalProperties: false
      description: Current and recent controlled monitoring exercises.
      properties:
        active:
          anyOf:
          - $ref: '#/components/schemas/OperationalExerciseResponse'
          - type: 'null'
        enabled:
          title: Enabled
          type: boolean
        recent:
          items:
            $ref: '#/components/schemas/OperationalExerciseResponse'
          title: Recent
          type: array
        supported_types:
          items:
            enum:
            - api_unavailable
            - high_latency
            - elevated_5xx
            type: string
          title: Supported Types
          type: array
      required:
      - enabled
      - active
      - recent
      - supported_types
      title: OperationalExerciseStateResponse
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: Generated from the deployed FastAPI routes and Pydantic models. Authenticated customer
    operations use Better Auth sessions or bearer tokens; internal and email-gateway routes require their
    documented service credentials.
  summary: Data, application, and integration contracts for Sicurre.
  title: Sicurre Platform API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/admin/operational-exercises:
    get:
      description: Return active state and recent audit records for platform administrators.
      operationId: get_operational_exercises_v1_admin_operational_exercises_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalExerciseStateResponse'
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Get Operational Exercises
      tags:
      - app-ui-flows
    post:
      description: Start one admin-only synthetic signal without affecting customer traffic.
      operationId: start_operational_exercise_v1_admin_operational_exercises_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationalExerciseCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalExerciseResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Start Operational Exercise
      tags:
      - app-ui-flows
  /v1/admin/operational-exercises/{exercise_id}/recover:
    post:
      description: Recover an active exercise early and preserve its audit trail.
      operationId: recover_operational_exercise_v1_admin_operational_exercises__exercise_id__recover_post
      parameters:
      - in: path
        name: exercise_id
        required: true
        schema:
          title: Exercise Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalExerciseResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Recover Operational Exercise
      tags:
      - app-ui-flows
...
servers:
- description: Current Sicurre deployment
  url: /
...